+++ zmscitizenapi/src/Zmscitizenapi/Services/Appointment/AppointmentByIdService.php
@@ -5,17 +5,17 @@ declare(strict_types=1);
 namespace BO\Zmscitizenapi\Services\Appointment;
 
 use BO\Zmscitizenapi\Models\ThinnedProcess;
-use BO\Zmscitizenapi\Models\Captcha\AltchaCaptcha;
+use BO\Zmscitizenapi\Services\Captcha\CaptchaService;
 use BO\Zmscitizenapi\Services\Core\ValidationService;
 use BO\Zmscitizenapi\Services\Core\ZmsApiFacadeService;
 
 class AppointmentByIdService
 {
-    private AltchaCaptcha $captchaService;
+    private CaptchaService $captchaService;
 
     public function __construct()
     {
-        $this->captchaService = new AltchaCaptcha();
+        $this->captchaService = new CaptchaService();
     }
 
     public function getAppointmentById(array $queryParams): ThinnedProcess|array
