+++ zmscitizenapi/tests/Zmscitizenapi/Controllers/Captcha/CaptchaVerifyControllerTest.php
@@ -3,7 +3,7 @@
 namespace BO\Zmscitizenapi\Tests\Controllers\Captcha;
 
 use BO\Zmscitizenapi\Tests\ControllerTestCase;
-use BO\Zmscitizenapi\Models\Captcha\AltchaCaptcha;
+use BO\Zmscitizenapi\Services\Captcha\CaptchaService;
 use GuzzleHttp\Handler\MockHandler;
 use GuzzleHttp\Psr7\Response;
 use GuzzleHttp\Client;
@@ -72,7 +72,7 @@ class CaptchaVerifyControllerTest extends ControllerTestCase
         $handlerStack = HandlerStack::create($mockHandler);
         $mockClient = new Client(['handler' => $handlerStack]);
 
-        $captcha = new class($mockClient) extends AltchaCaptcha {
+        $captcha = new class($mockClient) extends CaptchaService {
             public function __construct($client)
             {
                 parent::__construct();
@@ -104,7 +104,7 @@ class CaptchaVerifyControllerTest extends ControllerTestCase
         $handlerStack = HandlerStack::create($mockHandler);
         $mockClient = new Client(['handler' => $handlerStack]);
 
-        $captcha = new class($mockClient) extends AltchaCaptcha {
+        $captcha = new class($mockClient) extends CaptchaService {
             public function __construct($client)
             {
                 parent::__construct();
@@ -138,7 +138,7 @@ class CaptchaVerifyControllerTest extends ControllerTestCase
         $handlerStack = HandlerStack::create($mockHandler);
         $mockClient = new Client(['handler' => $handlerStack]);
 
-        $captcha = new class($mockClient) extends AltchaCaptcha {
+        $captcha = new class($mockClient) extends CaptchaService {
             public function __construct($client)
             {
                 parent::__construct();
@@ -169,7 +169,7 @@ class CaptchaVerifyControllerTest extends ControllerTestCase
         ]);
         $mockClient = new Client(['handler' => HandlerStack::create($mockHandler)]);
 
-        $captcha = new class($mockClient) extends AltchaCaptcha {
+        $captcha = new class($mockClient) extends CaptchaService {
             public function __construct($client)
             {
                 parent::__construct();
